Create Code File
Try to understand the logic. Everything is basic, In this Project User will first input the difficulty level then tell the range and then system will compare the user value by memory value and if the guess is correct, it will show success message and if it isnt it will show failure message. for generating a random value i have used a RandomRange function in which you give the range and it will generate the random number in eax register.